Issue #8 | Present raw values as integers

+++ bitalino.py
@@ -449,7 +449,7 @@ class BITalino(object):
             else:
                 number_bytes = int(math.ceil((52.+6.*(nChannels-4))/8.))
             
-            dataAcquired = numpy.zeros((nSamples, 5 + nChannels))
+            dataAcquired = numpy.zeros((nSamples, 5 + nChannels), dtype=int)
             for sample in range(nSamples):
                 Data = self.receive(number_bytes)
                 decodedData = list(struct.unpack(number_bytes * "B ", Data))
